Move pagination hook to hooks folder
